requirements.txt
fastapi
uvicorn
sqlmodel
passlib
bcrypt==4.0.1
pyjwt
python-multipart
psycopg2